Bigger column for file size

+++ database/migrations/2021_11_19_214727_file_size.php
@@ -14,7 +14,7 @@ class FileSize extends Migration
     public function up()
     {
         Schema::table('files', function (Blueprint $table) {
-            $table->unsignedInteger('size')->after('mime');
+            $table->unsignedBigInteger('size')->after('mime');
         });
     }
